+++ src/files/glob-source.js
@@ -6,20 +6,20 @@ const Path = require('path')
 const errCode = require('err-code')
 
 /**
-* Create an async iterator that yields paths that match requested file paths.
-*
-* @param {Iterable|AsyncIterable|String} paths File system path(s) to glob from
-* @param {Object} [options] Optional options
-* @param {Boolean} [options.recursive] Recursively glob all paths in directories
-* @param {Boolean} [options.hidden] Include .dot files in matched paths
-* @param {Array<String>} [options.ignore] Glob paths to ignore
-* @param {Boolean} [options.followSymlinks] follow symlinks
-* @param {Boolean} [options.preserveMode] preserve mode
-* @param {Boolean} [options.preserveMtime] preserve mtime
-* @param {Boolean} [options.mode] mode to use - if preserveMode is true this will be ignored
-* @param {Boolean} [options.mtime] mtime to use - if preserveMtime is true this will be ignored
-* @yields {Object} File objects in the form `{ path: String, content: AsyncIterator<Buffer> }`
-*/
+ * Create an async iterator that yields paths that match requested file paths.
+ *
+ * @param {Iterable<string> | AsyncIterable<string> | string} paths - File system path(s) to glob from
+ * @param {Object} [options] - Optional options
+ * @param {boolean} [options.recursive] - Recursively glob all paths in directories
+ * @param {boolean} [options.hidden] - Include .dot files in matched paths
+ * @param {Array<string>} [options.ignore] - Glob paths to ignore
+ * @param {boolean} [options.followSymlinks] - follow symlinks
+ * @param {boolean} [options.preserveMode] - preserve mode
+ * @param {boolean} [options.preserveMtime] - preserve mtime
+ * @param {number} [options.mode] - mode to use - if preserveMode is true this will be ignored
+ * @param {Date} [options.mtime] - mtime to use - if preserveMtime is true this will be ignored
+ * @yields {Object} File objects in the form `{ path: String, content: AsyncIterator<Buffer> }`
+ */
 module.exports = async function * globSource (paths, options) {
   options = options || {}
 
@@ -53,12 +53,14 @@ module.exports = async function * globSource (paths, options) {
     let mode = options.mode
 
     if (options.preserveMode) {
+      // @ts-ignore
       mode = stat.mode
     }
 
     let mtime = options.mtime
 
     if (options.preserveMtime) {
+      // @ts-ignore
       mtime = stat.mtime
     }
 
@@ -82,6 +84,7 @@ module.exports = async function * globSource (paths, options) {
   }
 }
 
+// @ts-ignore
 async function * toGlobSource ({ path, type, prefix, mode, mtime, preserveMode, preserveMtime }, options) {
   options = options || {}
 
@@ -135,4 +138,7 @@ async function * toGlobSource ({ path, type, prefix, mode, mtime, preserveMode, 
   }
 }
 
+/**
+ * @param {string} path
+ */
 const toPosix = path => path.replace(/\\/g, '/')