+++ src/files/url-source.js
@@ -1,7 +1,15 @@
 'use strict'
 
-const Http = require('../http')
+const HTTP = require('../http')
 
+/**
+ * @param {string} url
+ * @param {import("../types").HTTPOptions} [options]
+ * @returns {{
+    path: string;
+    content?: AsyncIterable<Uint8Array>;
+}}
+ */
 const urlSource = (url, options) => {
   return {
     path: decodeURIComponent(new URL(url).pathname.split('/').pop() || ''),
@@ -10,8 +18,9 @@ const urlSource = (url, options) => {
 }
 
 const readURLContent = async function * (url, options) {
-  const http = new Http()
+  const http = new HTTP()
   const response = await http.get(url, options)
+
   yield * response.iterator()
 }