+++ src/supports.js
@@ -1,9 +1,8 @@
 'use strict'
 
-const globalThis = require('./globalthis')
-
 module.exports = {
   supportsFileReader: typeof self !== 'undefined' && 'FileReader' in self,
   supportsWebRTC: 'RTCPeerConnection' in globalThis &&
-  (typeof navigator !== 'undefined' && typeof navigator.mediaDevices !== 'undefined' && 'getUserMedia' in navigator.mediaDevices)
+  (typeof navigator !== 'undefined' && typeof navigator.mediaDevices !== 'undefined' && 'getUserMedia' in navigator.mediaDevices),
+  supportsWebRTCDataChannels: 'RTCPeerConnection' in globalThis
 }